state machine cleanup
updated state machine to expect an AnimatedSprite2d with an animation

on the object's sprite2d - set canvasitem - visible = unchecked
(effectively we never see the sprite only its animations)

each child state in the state machine now requires a string in the Animation Name field of the animation to load from the AnimatedSprite2D file.

state_machine.gd tracks previous state, but is up to nodes to access it

added a 2 frame dummy animation to the aggro state for visual representation.

updated states to use programmatically generated timer instead of node based approach.  adds lines of code for tradeoff of removing scene nodes.  comments added with topics to investigate.